fix: extensive bug audit — 30+ fixes across C layer, Swift, and ChOma
ChOma/Mach-O:
- SHA-384 code slots now use CC_SHA384 (not CC_SHA256)
- Code directory verify_code_slot passes correct pageHash buffer
- DER entitlements tagged with correct slot type (CSSLOT_DER_ENTITLEMENTS)
- File/Memory stream auto-expand guard uses & not |
- File stream expand loop now actually runs (size_t i < n)
- Fat slice trim underflow guarded
- Superblob decode bounds-checks index offsets
- Superblob encode guards against integer overflow
- Mach-O load command cmdsize validated against overflow/zero/negative
- Mach-O ULEB128 reader bounds-checks before read
- Mach-O trie frontier realloc NULL-deref fixed
- Mach-O fileset entry_id offset validated before strdup

C memory safety:
- Stack buffer overflow in page-scan proc name (40→33 bytes)
- proc_get_p_name dangling pointer via temporary struct lifetime (strcmp on freed stack)
- VFS getrootvnode null-checks procbypid(1) result before krw
- vn_vname now PAC-strips v_name pointer for arm64e
- clearguardexc returns early if thread==0 instead of writing kernel address 0
- RemoteCall init reordered to validate port kobjects before kernel reads

Swift concurrency/resources:
- startPipeline guarded against re-entrance (prevents concurrent ds_run → kernel panic)
- LogManager deinit serialized on queue (race with concurrent append)
- PersistenceService load/save now use lock for external callers
- PersistenceService write hoisted outside loop (was writing per-app)
- IPAInstaller task cancelled on stream consumer disconnection
- IPAParser cleans up temp directory on all parse failure paths
- IPA temp file copies cleaned up after install
- Settings uicache error no longer silently swallowed (try? → do/catch)
- XPF offset storage uses Int64 not Int32 (no truncation on large values)

UI:
- FileBrowser empty directory correctly shows empty state (not VFS-not-ready error)
- FileBrowser isDirectory heuristic handles extensionless files
- ExploitProgressView stage counter excludes .failed from timeline

+++ Core/Kernel/XPFWrapper.swift
@@ -161,7 +161,7 @@ public enum XPFWrapper {
             }
 
             if oldValue != UInt32(resolved) {
-                defaults.set(Int32(resolved), forKey: key)
+                defaults.set(Int64(resolved), forKey: key)
                 overriddenCount += 1
                 LogManager.shared.append(
                     "XPF override: \(sym) → 0x\(String(resolved, radix: 16)) (was 0x\(String(oldValue, radix: 16)))",
